task-manager_no_template/pyproject.toml
[tool.poetry]
name = "fastapi-task-manager"
version = "0.1.0"
description = ""
authors = ["sergey <sereny9@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.40"
pydantic = "^2.11.3"
python-dotenv = "^1.1.0"
typing_extensions = "^4.13.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"